+++ ICSharpCode.ILSpyX/Settings/MutexProtector.cs
@@ -1,0 +1,55 @@
+// Copyright (c) 2026 AlphaSierraPapa for the SharpDevelop Team
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy of this
+// software and associated documentation files (the "Software"), to deal in the Software
+// without restriction, including without limitation the rights to use, copy, modify, merge,
+// publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons
+// to whom the Software is furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all copies or
+// substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+// INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
+// PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE
+// FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
+// OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+// DEALINGS IN THE SOFTWARE.
+
+using System;
+using System.Threading;
+
+namespace ICSharpCode.ILSpyX.Settings
+{
+	/// <summary>
+	/// Guards a config-file read-modify-write with a named <see cref="Mutex"/> so that two ILSpy
+	/// instances editing the same sidecar (the settings XML, the bookmark list, ...) fold their
+	/// changes in turn instead of overwriting each other. Acquire it around the
+	/// read-update-write block and dispose it to release.
+	/// </summary>
+	public sealed class MutexProtector : IDisposable
+	{
+		readonly Mutex mutex;
+
+		public MutexProtector(string name)
+		{
+			this.mutex = new Mutex(true, name, out bool createdNew);
+			if (createdNew)
+				return;
+
+			try
+			{
+				mutex.WaitOne();
+			}
+			catch (AbandonedMutexException)
+			{
+			}
+		}
+
+		public void Dispose()
+		{
+			mutex.ReleaseMutex();
+			mutex.Dispose();
+		}
+	}
+}
